fix(v2): weeklyDelta Int → Double (was crashing all of Home in 146)
The backend (7b50b6a) started returning weeklyDelta as a one-decimal
Double ('1.8 tasks/wk') but iOS still decoded it as Int. Codable
silently rejected the whole V2HomeData payload, vm.data stayed nil,
and Home rendered the 'Couldn't load your day. Pull to refresh.'
error state for EVERY user. The 'Set your objective' pill is just
the default-when-nil string — not a real no-objective state.

Same fix applied to V2CalibrationInsightsViewModel's Trajectory
struct (identical shape, would have crashed the same way the moment
a user re-opened calibration).

Build 147 ships the type fix.

+++ ScaleUp/Features/V2/Diagnostic/V2CalibrationInsightsViewModel.swift
@@ -64,7 +64,7 @@ struct V2InsightsData: Codable {
         let atTargetDate: Int
         let targetReadiness: Int
         let timelineWeeks: Int
-        let weeklyDelta: Int
+        let weeklyDelta: Double
         let onTrack: Bool
         let headline: String?
     }
